+++ rbtree_augmented.h
@@ -87,7 +87,7 @@ rbstatic const struct rb_augment_callbacks rbname = {			\
 #define	RB_RED		0
 #define	RB_BLACK	1
 
-#define __rb_parent(pc)    ((struct rb_node *)(pc & ~3))
+#define __rb_parent(pc)    ((struct rb_node *)(intptr_t)(pc & ~3))
 
 #define __rb_color(pc)     ((pc) & 1)
 #define __rb_is_black(pc)  __rb_color(pc)
@@ -98,13 +98,13 @@ rbstatic const struct rb_augment_callbacks rbname = {			\
 
 static /*inline*/ void rb_set_parent(struct rb_node *rb, struct rb_node *p)
 {
-	rb->__rb_parent_color = rb_color(rb) | (unsigned long)p;
+	rb->__rb_parent_color = rb_color(rb) | (unsigned long)(intptr_t)p;
 }
 
 static /*inline*/ void rb_set_parent_color(struct rb_node *rb,
 				       struct rb_node *p, int color)
 {
-	rb->__rb_parent_color = (unsigned long)p | color;
+	rb->__rb_parent_color = (unsigned long)(intptr_t)p | color;
 }
 
 static /*inline*/ void
